refactor: clean up ledger onboarding files
