+++ packages/shared/features/onboarding.features.ts
@@ -144,10 +144,10 @@ const onboardingFeatures = {
             newProfile: {
                 enabled: false,
                 softwareProfile: {
-                    enabled: false,
+                    enabled: true,
                 },
                 ledgerProfile: {
-                    enabled: false,
+                    enabled: true,
                 },
             },
             restoreProfile: {
@@ -188,7 +188,7 @@ const onboardingFeatures = {
                     enabled: true,
                 },
                 ledgerProfile: {
-                    enabled: false,
+                    enabled: true,
                 },
             },
             restoreProfile: {
@@ -204,7 +204,7 @@ const onboardingFeatures = {
                     enabled: true,
                 },
                 ledgerBackup: {
-                    enabled: false,
+                    enabled: true,
                 },
             },
         },
