+++ packages/shared/routes/onboarding/views/ledger-setup/views/index.js
@@ -1,6 +1,2 @@
-export { default as AccountIndexView } from './AccountIndexView.svelte'
-export { default as ConnectView } from './ConnectView.svelte'
-export { default as GenerateNewAddressView } from './GenerateNewAddressView.svelte'
-export { default as InstallationGuideView } from './InstallationGuideView.svelte'
-export { default as RestoreFromLedgerView } from './RestoreFromLedgerView.svelte'
-export { default as SwitchAppsView } from './SwitchAppsView.svelte'
+export { default as ConnectLedgerView } from './ConnectLedgerView.svelte'
+export { default as LedgerInstallationGuideView } from './LedgerInstallationGuideView.svelte'
